refactor: Encryptor to gnr.core.gnrcrypto, encryption_key on db
- Move Encryptor from gnr.sql.gnrsql.encryption to gnr.core.gnrcrypto
  (reusable outside SQL layer)
- Remove gnr.sql.gnrsql.encryption module
- encryption_key as db attribute (like dbname, host, password)
- Lazy encryptor property on GnrSqlDb, always returns an Encryptor
  (raises on use if key or cryptography missing, not on access)
- Remove encryptor from GnrApp and GnrSqlAppDb
- All SQL layer uses self.db.encryptor directly, no getattr
- Remove pantry dependency from db.py (check is in Encryptor itself)
- Simplified all encrypt/decrypt call sites

Refs #747
