+++ gnrpy/gnr/sql/adapters/_gnrbaseadapter.py
@@ -736,8 +736,6 @@ class SqlDbAdapter(object):
         """
         data_out = {}
         tbl_virtual_columns = tblobj.virtual_columns
-        app = getattr(self.dbroot, 'application', None)
-        encryptor = getattr(app, 'encryptor', None) if app else None
         for k in list(record_data.keys()):
             if not (k.startswith('@') or k=='pkey' or  k in tbl_virtual_columns):
                 v = record_data[k]
@@ -748,12 +746,11 @@ class SqlDbAdapter(object):
                                 innernode.attr.pop(blackattr,None)
                     v = v.toXml() if v else None
                     #data_out[str(k.lower())] = v
-                if encryptor and v is not None:
-                    col = tblobj.columns.get(k)
-                    if col:
-                        enc_mode = col.attributes.get('encrypted')
-                        if enc_mode:
-                            v = encryptor.encrypt(v, enc_mode)
+                col = tblobj.columns.get(k)
+                if col and v is not None:
+                    enc_mode = col.attributes.get('encrypted')
+                    if enc_mode:
+                        v = self.dbroot.encryptor.encrypt(v, enc_mode)
                 data_out[str(k)] = v
         sql_value_cols = [k for k,v in list(tblobj.columns.items()) if 'sql_value' in v.attributes and not k in data_out]
         for k in sql_value_cols:
@@ -1410,13 +1407,11 @@ class GnrWhereTranslator(object):
                     value = sqlArgs.pop(attr['value_caption'],'')
                 enc_mode = attr.get('encrypted') or colobj.attributes.get('encrypted')
                 if enc_mode == 'Q' and value is not None:
-                    app = getattr(self.db, 'application', None)
-                    encryptor = getattr(app, 'encryptor', None) if app else None
-                    if encryptor:
-                        if isinstance(value, list):
-                            value = [encryptor.encrypt(v, 'Q') for v in value]
-                        else:
-                            value = encryptor.encrypt(value, 'Q')
+                    encryptor = self.db.encryptor
+                    if isinstance(value, list):
+                        value = [encryptor.encrypt(v, 'Q') for v in value]
+                    else:
+                        value = encryptor.encrypt(value, 'Q')
                 onecondition = self.prepareCondition(column, op, value, dtype, sqlArgs,tblobj=tblobj,parname=parname)
 
             if onecondition:
